Namespace for data transmission tests
