+++ .github/workflows/all_solutions.yml
@@ -444,7 +444,7 @@ jobs:
     strategy:
       matrix:
         namespace: [ AgentFeatures, Api, AspNetCore, BasicInstrumentation, CatInbound, CatOutbound, CSP,
-          CustomAttributes, CustomInstrumentation, DistributedTracing, DistributedTracing.W3CInstrumentationTests, 
+          CustomAttributes, CustomInstrumentation, DataTransmission, DistributedTracing, 
           HttpClientInstrumentation.NetCore, HttpClientInstrumentation.NetFramework, Logging, Owin, ReJit.NetCore, 
           ReJit.NetFramework, RemoteServiceFixtures, RestSharp, WCF.Client.IIS.ASPDisabled, WCF.Client.IIS.ASPEnabled,
           WCF.Client.Self, WCF.Service.IIS.ASPDisabled, WCF.Service.IIS.ASPEnabled, WCF.Service.Self ]
